# Proof of concept: skip unused Gitignore matchers