This is just a proof of concept. I'm not sure if this is the best approach. Even if it is, the code should probably be clean up.

This PR addresses the findings in #835 and disables unneeded `Gitignore` matchers completely. This leads to a performance improvement for use cases where one or more of the matchers (custom ignore files, `.ignore` files, `.gitignore` files, `.git/info/exclude`) are not needed.



### Benchmark 1: `rg -uu --files`

master
```
▶ hyperfine --warmup 10 'rg -uu --files'
Benchmark #1: rg -uu --files

  Time (mean ± σ):      1.298 s ±  0.067 s    [User: 4.348 s, System: 2.667 s]
 
  Range (min … max):    1.203 s …  1.398 s
```

this branch (~ 8% faster):
```
▶ hyperfine --warmup 10 'rg -uu --files'
Benchmark #1: rg -uu --files

  Time (mean ± σ):      1.201 s ±  0.032 s    [User: 3.732 s, System: 2.532 s]
 
  Range (min … max):    1.165 s …  1.264 s
```

### Benchmark 2: `rg -uu --files --glob "Cargo.toml"`

master:
```
▶ hyperfine --warmup 10 'rg -uu --files --glob "Cargo.toml"'
Benchmark #1: rg -uu --files --glob "Cargo.toml"

  Time (mean ± σ):     662.5 ms ±   2.2 ms    [User: 2.991 s, System: 2.198 s]
 
  Range (min … max):   659.3 ms … 665.7 ms
```

this branch (~ 10% faster):
```
▶ hyperfine --warmup 10 'rg -uu --files --glob "Cargo.toml"' 
Benchmark #1: rg -uu --files --glob "Cargo.toml"

  Time (mean ± σ):     602.6 ms ±   1.7 ms    [User: 2.527 s, System: 2.183 s]
 
  Range (min … max):   599.1 ms … 604.8 ms
```

### Benchmark 3: `rg --files --glob "Cargo.toml"`

As expected, there is no significant change if `-uu` is not used:

master:
```
▶ hyperfine --warmup 10 'rg --files --glob "Cargo.toml"'    
Benchmark #1: rg --files --glob "Cargo.toml"

  Time (mean ± σ):     101.1 ms ±   1.6 ms    [User: 420.4 ms, System: 327.0 ms]
 
  Range (min … max):    99.5 ms … 107.1 ms
```

this branch:
```
▶ hyperfine --warmup 10 'rg --files --glob "Cargo.toml"'
Benchmark #1: rg --files --glob "Cargo.toml"

  Time (mean ± σ):     101.6 ms ±   0.6 ms    [User: 419.2 ms, System: 331.5 ms]
 
  Range (min … max):   100.3 ms … 103.0 ms
```