@BurntSushi 2018-03-10
@sharkdp Thanks for this! I think I'd be OK with this PR, but just to cover our bases, did you explore whether this improvement was possible by just using `Gitignore::empty()`? That was basically my intent, where if a matcher wasn't needed, then we could create an empty one and the empty cases would just get folded into the matcher instead of needing to add a bunch of explicit case analysis everywhere.

Let's work through this. Here's the definition of `Gitignore::empty()`:

```rust
    pub fn empty() -> Gitignore {
        GitignoreBuilder::new("").build().unwrap()
    }
```

OK, so this is creating a new builder with an empty root, not adding any globs and then building it. That means running this code:

```rust
    pub fn new<P: AsRef<Path>>(root: P) -> GitignoreBuilder {
        let root = root.as_ref();
        GitignoreBuilder {
            builder: GlobSetBuilder::new(),
            root: strip_prefix("./", root).unwrap_or(root).to_path_buf(),
            globs: vec![],
            case_insensitive: false,
        }
    }
```

and this code

```rust
    pub fn build(&self) -> Result<Gitignore, Error> {
        let nignore = self.globs.iter().filter(|g| !g.is_whitelist()).count();
        let nwhite = self.globs.iter().filter(|g| g.is_whitelist()).count();
        let set =
            self.builder.build().map_err(|err| {
                Error::Glob {
                    glob: None,
                    err: err.to_string(),
                }
            })?;
        Ok(Gitignore {
            set: set,
            root: self.root.clone(),
            globs: self.globs.clone(),
            num_ignores: nignore as u64,
            num_whitelists: nwhite as u64,
            matches: Arc::new(ThreadLocal::default()),
        })
    }
```

The constructor seems pretty cheap to me. There's a call to `to_path_buf` which might invoke an allocation, although given it's an empty string, that probably gets  optimized out. (IIRC, `vec![]` doesn't actually allocate, for example, and `PathBuf` is certainly built on `Vec`.)

The `build` method is a little more interesting. The first couple lines filter over the globs, but we have none, so that should be negligible. Then we build the `GlobSet` itself. This is actually cheap since `GlobSet::new(&[])` is special cases to do no work and just return a set that never matches.

The builder otherwise does some cloning (but again, those should all be empty), and sets up a scratch buffer (which also defaults to empty).

So my question is: if we changed this

```rust
        let custom_ig_matcher =
            {
                let (m, err) =
                    create_gitignore(&dir, &self.0.custom_ignore_filenames);
                errs.maybe_push(err);
                m
            };
```

to be more in line with how the other matchers handle the empty case on master

```rust
        let custom_ig_matcher =
            if self.custom_ignore_filenames.is_empty() {
                Gitignore::empty()
            } else {
                let (m, err) =
                    create_gitignore(&dir, &self.0.custom_ignore_filenames);
                errs.maybe_push(err);
                m
            };
```

then does that solve the performance problem? Or are there costs in `Gitignore::empty()` that I am not accounting for?

(The reason why I think the above might work is because the `create_gitignore` function definitely does have overhead associated with it, and always does at least a syscall to open a file.