@sharkdp 2018-03-12
@BurntSushi Thank you for the detailed answer.

I should have explained this better:

1. I'm pretty sure that the problem is *not* with the additional `Gitignore::empty` invocation. The expensive operation is the `matched(&path, is_dir)` call on the four `Gitignore` instances in (every loop iteration inside) `Ignore::matched_ignore`.

2. This PR does not only resolve the performance regression in #835 but improves the speed beyond that (if more than one matcher is disabled).